pyproject.toml
[tool.poetry]
name = "backend-fractal"
version = "0.1.0"
description = "Fractal Backend Django Challenge"
authors = ["Júlio Agripino <juliocesar014@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
django-ninja = "^1.3.0"
pydantic = "^2.10.0"
email-validator = "^2.2.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2 = "^2.9.10"
python-memcached = "^1.62"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runserver = "django.core.management:execute_from_command_line"
